refactor(proxy): split past meeting and participant client methods
Move past meeting, summary, invitee, and attendee ITX calls into files
aligned with the domain past-meeting and participant client interfaces.
