+++ internal/infrastructure/proxy/client_api.go
@@ -10,130 +10,6 @@ import (
 	"github.com/linuxfoundation/lfx-v2-meeting-service/pkg/models/itx"
 )
 
-func (c *Client) CreatePastMeeting(ctx context.Context, req *itx.CreatePastMeetingRequest) (*itx.PastMeetingResponse, error) {
-	return doJSONTyped[itx.PastMeetingResponse](c, ctx, apiRequest{
-		method:     http.MethodPost,
-		path:       "/v2/zoom/past_meetings",
-		body:       req,
-		accept:     acceptJSON,
-		parseError: "failed to unmarshal response",
-	})
-}
-
-// GetPastMeeting retrieves a past meeting record via ITX proxy.
-func (c *Client) GetPastMeeting(ctx context.Context, pastMeetingID string) (*itx.PastMeetingResponse, error) {
-	return doJSONTyped[itx.PastMeetingResponse](c, ctx, apiRequest{
-		method:     http.MethodGet,
-		path:       "/v2/zoom/past_meetings/%s",
-		pathArgs:   []any{pastMeetingID},
-		accept:     acceptJSON,
-		parseError: "failed to unmarshal response",
-	})
-}
-
-// UpdatePastMeeting updates a past meeting record via ITX proxy.
-// Returns nil on success (ITX API returns 204 No Content).
-func (c *Client) UpdatePastMeeting(ctx context.Context, pastMeetingID string, req *itx.CreatePastMeetingRequest) (*itx.PastMeetingResponse, error) {
-	return doJSONTypedOptional[itx.PastMeetingResponse](c, ctx, apiRequest{
-		method:   http.MethodPut,
-		path:     "/v2/zoom/past_meetings/%s",
-		pathArgs: []any{pastMeetingID},
-		body:     req,
-		accept:   acceptJSON,
-	})
-}
-
-// DeletePastMeeting deletes a past meeting record via ITX proxy.
-func (c *Client) DeletePastMeeting(ctx context.Context, pastMeetingID string) error {
-	return c.doNoContent(ctx, apiRequest{
-		method:   http.MethodDelete,
-		path:     "/v2/zoom/past_meetings/%s",
-		pathArgs: []any{pastMeetingID},
-	})
-}
-
-// GetPastMeetingSummary retrieves a past meeting summary from ITX.
-func (c *Client) GetPastMeetingSummary(ctx context.Context, pastMeetingID, summaryID string) (*itx.PastMeetingSummaryResponse, error) {
-	return doJSONTyped[itx.PastMeetingSummaryResponse](c, ctx, apiRequest{
-		method:     http.MethodGet,
-		path:       "/v2/zoom/past_meetings/%s/summaries/%s",
-		pathArgs:   []any{pastMeetingID, summaryID},
-		accept:     acceptJSON,
-		parseError: "failed to unmarshal response",
-	})
-}
-
-// UpdatePastMeetingSummary updates a past meeting summary in ITX.
-func (c *Client) UpdatePastMeetingSummary(ctx context.Context, pastMeetingID, summaryID string, req *itx.UpdatePastMeetingSummaryRequest) (*itx.PastMeetingSummaryResponse, error) {
-	return doJSONTyped[itx.PastMeetingSummaryResponse](c, ctx, apiRequest{
-		method:     http.MethodPut,
-		path:       "/v2/zoom/past_meetings/%s/summaries/%s",
-		pathArgs:   []any{pastMeetingID, summaryID},
-		body:       req,
-		accept:     acceptJSON,
-		parseError: "failed to unmarshal response",
-	})
-}
-
-func (c *Client) CreateInvitee(ctx context.Context, pastMeetingID string, req *itx.CreateInviteeRequest) (*itx.InviteeResponse, error) {
-	return doJSONTyped[itx.InviteeResponse](c, ctx, apiRequest{
-		method:     http.MethodPost,
-		path:       "/v2/zoom/past_meetings/%s/invitees",
-		pathArgs:   []any{pastMeetingID},
-		body:       req,
-		parseError: "failed to unmarshal response",
-	})
-}
-
-// UpdateInvitee updates an invitee for a past meeting via the ITX proxy.
-func (c *Client) UpdateInvitee(ctx context.Context, pastMeetingID, inviteeID string, req *itx.UpdateInviteeRequest) (*itx.InviteeResponse, error) {
-	return doJSONTypedOptional[itx.InviteeResponse](c, ctx, apiRequest{
-		method:   http.MethodPut,
-		path:     "/v2/zoom/past_meetings/%s/invitees/%s",
-		pathArgs: []any{pastMeetingID, inviteeID},
-		body:     req,
-	})
-}
-
-// DeleteInvitee deletes an invitee from a past meeting via the ITX proxy.
-func (c *Client) DeleteInvitee(ctx context.Context, pastMeetingID, inviteeID string) error {
-	return c.doNoContent(ctx, apiRequest{
-		method:   http.MethodDelete,
-		path:     "/v2/zoom/past_meetings/%s/invitees/%s",
-		pathArgs: []any{pastMeetingID, inviteeID},
-	})
-}
-
-// CreateAttendee creates an attendee for a past meeting via the ITX proxy.
-func (c *Client) CreateAttendee(ctx context.Context, pastMeetingID string, req *itx.CreateAttendeeRequest) (*itx.AttendeeResponse, error) {
-	return doJSONTyped[itx.AttendeeResponse](c, ctx, apiRequest{
-		method:     http.MethodPost,
-		path:       "/v2/zoom/past_meetings/%s/attendees",
-		pathArgs:   []any{pastMeetingID},
-		body:       req,
-		parseError: "failed to unmarshal response",
-	})
-}
-
-// UpdateAttendee updates an attendee for a past meeting via the ITX proxy.
-func (c *Client) UpdateAttendee(ctx context.Context, pastMeetingID, attendeeID string, req *itx.UpdateAttendeeRequest) (*itx.AttendeeResponse, error) {
-	return doJSONTypedOptional[itx.AttendeeResponse](c, ctx, apiRequest{
-		method:   http.MethodPut,
-		path:     "/v2/zoom/past_meetings/%s/attendees/%s",
-		pathArgs: []any{pastMeetingID, attendeeID},
-		body:     req,
-	})
-}
-
-// DeleteAttendee deletes an attendee from a past meeting via the ITX proxy.
-func (c *Client) DeleteAttendee(ctx context.Context, pastMeetingID, attendeeID string) error {
-	return c.doNoContent(ctx, apiRequest{
-		method:   http.MethodDelete,
-		path:     "/v2/zoom/past_meetings/%s/attendees/%s",
-		pathArgs: []any{pastMeetingID, attendeeID},
-	})
-}
-
 func (c *Client) CreateMeetingAttachmentPresignURL(ctx context.Context, meetingID string, req *itx.CreateAttachmentPresignRequest) (*itx.MeetingAttachmentPresignResponse, error) {
 	return doJSONTyped[itx.MeetingAttachmentPresignResponse](c, ctx, apiRequest{
 		method:      http.MethodPost,
